requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pydantic==1.10.14
python-multipart==0.0.9
pillow==10.3.0
pdf2image==1.17.0
pytesseract==0.3.10
paddlepaddle==2.5.2
paddleocr==2.7.0.3
python-docx==1.1.0
numpy==1.26.4
